requirements.txt.txt
folium==0.13.0
Pillow==9.2.0
wordcloud==1.8.2.2
matplotlib==3.6.2
numpy==1.23.4
pandas==1.5.1
plotly==5.9.0
seaborn==0.12.1
streamlit==1.11.0
streamlit-folium==0.7.0